backend_api/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
openai-whisper==20240930
keras>=3.0.0
tensorflow>=2.15.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
librosa>=0.10.0
soundfile>=0.12.0
